deps.isdirty() folded into state.File.is_dirty()
It just seemed like a method of File, y'know?

One wrinkle is that is_checked() and set_checked[_save]() are sometimes
called as methods and sometimes as static methods (accessed though File
rather than self, which is perfectly legitimate python idiom).

This works fine now, but if anyone ever subclasses File and overrides
those methods the current incarnation of is_dirty won't use them
automatically.

It escapes me right now just exactly how to do a proper MRO lookup of a
method and then get it's unbound form so I'll leave it as it is for now.